codewhisperer-backend/requirements.txt
Flask>=2.0.0
flask-cors>=3.0.0
Flask-SQLAlchemy>=3.0.0
google-generativeai>=0.8.0
openai>=1.0.0
requests>=2.25.0
numpy>=1.20.0
scikit-learn>=1.0.0
scipy>=1.7.0
joblib>=1.0.0
tqdm>=4.60.0
pydantic>=2.0.0
